+++ src/OpenTelemetry.Sampler.AWS/CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ## Unreleased
 
-* Updated OpenTelemetry core component version(s) to `1.15.1`.
-  ([#4020](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4020))
+## 0.1.0-alpha.8
+
+Released 2026-Apr-14
+
+* Updated OpenTelemetry core component version(s) to `1.15.2`.
+  ([#4080](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4080))
+* Limit the max size read for response body getting the sampling rules to 1MB.
+  ([#4100](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4100))
 
 ## 0.1.0-alpha.7